requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
